refactor(settings): split 'superdesk-settings' module.
`superdesk-settings` becomes:
```
.
├── directives
│   ├── DateParam.js
│   ├── RoleUnique.js
│   ├── SettingsView.js
│   ├── ValidError.js
│   └── index.js
├── styles
│   └── settings.less
├── views
│   ├── main.html
│   └── settings-view.html
└── index.js

3 directories, 9 files
```
