+++ scripts/superdesk-settings/directives/ValidError.js
@@ -1,0 +1,7 @@
+export function ValidError() {
+    return {
+        link: function (scope, element) {
+            element.addClass('validation-error');
+        }
+    };
+}
